crud hecho

+++ model/camareroDAO.php
@@ -33,6 +33,14 @@ class camareroDAO {
         }
     }
 
+    public function getCamarero($id) {
+        $query = "SELECT * FROM camareros WHERE id_camarero = ?";
+        $sentencia = $this->pdo->prepare($query);
+        $sentencia->bindParam(1,$id);
+        $sentencia->execute();
+        return $sentencia->fetch(PDO::FETCH_ASSOC);
+    }
+
     public function readCamareros() {
         $query = "SELECT * FROM camareros;";
         $sentencia=$this->pdo->prepare($query);
@@ -40,11 +48,10 @@ class camareroDAO {
         $lista_camareros=$sentencia->fetchAll(PDO::FETCH_ASSOC);
         
         foreach($lista_camareros as $camarero) {
-
             $id=$camarero['id_camarero'];
             echo "<tr>";
             echo "<td style='border:1px solid black'><a href='../view/modificarCamarero.php?id_camarero=$id&'>Modificar</a></th>";
-            echo "<td style='border:1px solid black'><a href='../view/index.admin.php?eliminar=true&id_camarero=$id'>Eliminar</a></th>";
+            echo "<td style='border:1px solid black'><a href='../view/index.admin.php?eliminar=true&id_camarero=$id' OnClick='return confirm(`¿Estás seguro?`);'>Eliminar</a></th>";
             echo "<td style='border:1px solid black'>{$camarero['nombre_camarero']}</th>";
             echo "<td style='border:1px solid black'>{$camarero['pass_camarero']}</th>";
             if($camarero['rol'] == 0) {
@@ -56,11 +63,12 @@ class camareroDAO {
             }
             echo "</tr>";
         }
+
+        return $lista_camareros;
     }
 
     public function añadir() {
         try {
-            $this->pdo->beginTransaction();
             // RECOGEMOS LOS DATOS DEL NUEVO camareros
             $nombre=$_POST['nombre_camarero'];
             $pass=md5($_POST['contrasenya']);
@@ -74,46 +82,87 @@ class camareroDAO {
             $sentencia->bindParam(3,$rol);
             $sentencia->execute();
             
-            $this->pdo->commit();
-            header('Location: ../view/index.admin.php');
-            
         } catch (Exception $e) {
-            $this->pdo->rollBack();
             echo $e;
         }
     }
 
     public function eliminarCamarero($id) {
         try {
-            $pdo->beginTransaction();
+            $this->pdo->beginTransaction();
 
-            $query = "SELECT * FROM reserva WHERE `id_camarero` = ?";
-            $sentencia=$pdo->prepare($query);
-            $sentencia->bindParam(1,$id);
-            $sentencia->execute();
-            $lista_reservas=$sentencia->fetchAll(PDO::FETCH_ASSOC);
-
-            if($lista_notas!="") {    
-                $query="DELETE FROM `tbl_camareros` WHERE `id_camarero` = ?";
-                $sentencia=$pdo->prepare($query);
-                $sentencia->bindParam(1,$id);
-                $sentencia->execute();    
-            } else {
-                $query="DELETE FROM `tbl_notas` WHERE `id_camarero` = ?";
-                $sentencia=$pdo->prepare($query);
+            //Miramos que el usuario a eliminar no es ni el propio usuario ni el camarero uno (el cual es fijo y no se puede eliminar)
+            if($_SESSION['camarero']->getId_camarero() != $id && $id!=1) {
+                //Miramos si el usuario en cuestion tiene alguna reserva a su cargo
+                $query = "SELECT * FROM reserva WHERE `id_camarero` = ?";
+                $sentencia=$this->pdo->prepare($query);
                 $sentencia->bindParam(1,$id);
                 $sentencia->execute();
-            
-                $query="DELETE FROM `tbl_camareros` WHERE `id_camarero` = ?";
-                $sentencia=$pdo->prepare($query);
-                $sentencia->bindParam(1,$id);
-                $sentencia->execute();  
+                $lista_reservas=$sentencia->fetchAll(PDO::FETCH_ASSOC);
+                
+                //En caso de que tenga alguna reserva a su cargo,
+                //antes de eliminar el usuario, asignaremos sus reservas a otro camarero existente.
+                //En el caso de que no tenga ninguna reserva a su cargo, se elimina directamente.
+                if($lista_reservas!="") {    
+                    $query = "SELECT `id_camarero` FROM `camareros` WHERE `rol`='0'";
+                    $sentencia=$this->pdo->prepare($query);
+                    $sentencia->execute();
+                    $lista_camareros=$sentencia->fetchAll(PDO::FETCH_ASSOC);
+                    $idCam = $lista_camareros[1]['id_camarero'];
+                    
+                    if($idCam != "" && $idCam != $id) {
+                        $query = "UPDATE reserva SET id_camarero=? WHERE id_camarero=?";
+                        $sentencia=$this->pdo->prepare($query);
+                        $sentencia->bindParam(1,$idCam);
+                        $sentencia->bindParam(2,$id);
+                        $sentencia->execute();
+
+                        $query="DELETE FROM `camareros` WHERE `id_camarero` = ?";
+                        $sentencia=$this->pdo->prepare($query);
+                        $sentencia->bindParam(1,$id);
+                        $sentencia->execute();
+                    } else if($idCam != "" && $idCam == $id) {
+                        $query = "UPDATE reserva SET id_camarero=1 WHERE id_camarero=?";
+                        $sentencia=$this->pdo->prepare($query);
+                        $sentencia->bindParam(1,$id);
+                        $sentencia->execute();
+
+                        $query="DELETE FROM `camareros` WHERE `id_camarero` = ?";
+                        $sentencia=$this->pdo->prepare($query);
+                        $sentencia->bindParam(1,$id);
+                        $sentencia->execute();
+                    }
+                } else {
+                    $query="DELETE FROM `camareros` WHERE `id_camarero` = ?";
+                    $sentencia=$this->pdo->prepare($query);
+                    $sentencia->bindParam(1,$id);
+                    $sentencia->execute();
+                }
             }
             
-            $pdo->commit();
+            $this->pdo->commit();
             header('Location: ../view/index.admin.php');
         } catch (Exception $e) {
             $this->pdo->rollBack();
+        }
+    }
+
+    public function modificarCamarero() {
+        try {
+            $id = $_POST['id_camarero_modificar'];
+            $nombre = $_POST['nombre_camarero'];
+            $pass = md5($_POST['contrasenya']);
+            $rol = $_POST['rol'];
+
+            $query = "UPDATE camareros SET nombre_camarero=?, pass_camarero=?, rol=? WHERE id_camarero=?";
+            $sentencia=$this->pdo->prepare($query);
+            $sentencia->bindParam(1,$nombre);
+            $sentencia->bindParam(2,$pass);
+            $sentencia->bindParam(3,$rol);
+            $sentencia->bindParam(4,$id);
+            $sentencia->execute();
+            header('Location: ../view/index.admin.php');
+        } catch (Exception $e) {
             echo $e;
         }
     }
